+++ sqlite/common.go
@@ -10,18 +10,16 @@ import (
 )
 
 type EnvService struct {
-	db      *sql.DB
-	keyring domain.Keyring
+	db *sql.DB
 }
 
-func NewEnvService(ctx context.Context, dsn string, keyring domain.Keyring) (domain.EnvService, error) {
+func NewEnvService(ctx context.Context, dsn string) (domain.EnvService, error) {
 	// TODO use context!!
 	db, err := sqlite.Connect(dsn)
 	if err != nil {
 		return nil, fmt.Errorf("could not init db: %w", err)
 	}
 	return &EnvService{
-		db:      db,
-		keyring: keyring,
+		db: db,
 	}, nil
 }