+++ sqlite/sqlite/migrations/005_drop_keyring_entry.sql
@@ -1,0 +1,1 @@
+DROP TABLE keyring_entry;